Theia

+++ pkg/misc/subcmd/subcmd.go
@@ -8,7 +8,9 @@ import (
 	"syscall"
 	"time"
 
+	"github.com/keybase/go-ps"
 	"github.com/manifold/tractor/pkg/data/icons"
+	"github.com/manifold/tractor/pkg/misc/logging"
 )
 
 type Status string
@@ -18,6 +20,13 @@ const (
 	StatusStarted  Status = "Started"
 	StatusExited   Status = "Exited"
 	StatusStopped  Status = "Stopped"
+)
+
+var (
+	ErrStarted    = errors.New("already started")
+	ErrStarting   = errors.New("already starting")
+	ErrNotRunning = errors.New("not running")
+	ErrWaiting    = errors.New("already waiting")
 )
 
 func (s Status) Icon() []byte {
@@ -41,15 +50,19 @@ func Running(c *Subcmd) bool {
 	return c.status == StatusStarting || c.status == StatusStarted
 }
 
+type Observer func(*Subcmd, Status)
+
 type Subcmd struct {
 	*exec.Cmd
 
 	Setup       func(*exec.Cmd) error
-	MaxRestarts int
+	maxRestarts int
+
+	Log logging.DebugLogger
 
 	Started chan *exec.Cmd
 
-	callbacks []func(*Subcmd)
+	callbacks []Observer
 
 	current *exec.Cmd
 	status  Status
@@ -60,29 +73,36 @@ type Subcmd struct {
 
 	waitCh chan error
 
-	cbMu   sync.Mutex
-	runMu  sync.Mutex
-	statMu sync.Mutex
-	waitMu sync.Mutex
-	pidMu  sync.Mutex
+	cbMu      sync.Mutex
+	runMu     sync.Mutex
+	statMu    sync.Mutex
+	waitMu    sync.Mutex
+	pidMu     sync.Mutex
+	lastMu    sync.Mutex
+	restartMu sync.Mutex
 }
 
 func New(name string, arg ...string) *Subcmd {
-	s := &Subcmd{
+	return &Subcmd{
 		Cmd:         exec.Command(name, arg...),
-		MaxRestarts: -1,
+		maxRestarts: -1,
 		status:      StatusStopped,
 	}
-	return s
 }
 
-func (sc *Subcmd) OnStatusChange(cb func(*Subcmd)) {
-	cb(sc)
+func (sc *Subcmd) SetMaxRestarts(n int) {
+	sc.restartMu.Lock()
+	defer sc.restartMu.Unlock()
+	sc.maxRestarts = n
+}
+
+func (sc *Subcmd) Observe(cb Observer) {
 	sc.cbMu.Lock()
 	sc.callbacks = append(sc.callbacks, cb)
 	sc.cbMu.Unlock()
 }
 
+// don't call in observer callback!
 func (sc *Subcmd) Status() Status {
 	sc.statMu.Lock()
 	defer sc.statMu.Unlock()
@@ -91,44 +111,54 @@ func (sc *Subcmd) Status() Status {
 
 func (sc *Subcmd) Start() error {
 	if sc.Status() == StatusStarting || sc.Status() == StatusStarted {
-		return errors.New("already started")
+		return ErrStarted
 	}
 	return sc.start()
 }
 
 func (sc *Subcmd) Restart() error {
 	if sc.Status() == StatusStarting {
-		return errors.New("already starting")
+		return ErrStarting
 	}
 	if !Running(sc) {
 		return sc.start()
 	}
-	return sc.terminate(false)
+	return sc.terminate()
 }
 
 func (sc *Subcmd) Stop() error {
 	if !Running(sc) {
-		return errors.New("not running")
+		return ErrNotRunning
 	}
-	return sc.terminate(true)
+	sc.setStatus(StatusStopped)
+	return sc.terminate()
 }
 
-func (sc *Subcmd) terminate(stop bool) error {
-	if stop {
-		defer sc.setStatus(StatusStopped)
-	}
+func (sc *Subcmd) Signal(sig os.Signal) {
 	sc.pidMu.Lock()
+	sc.current.Process.Signal(sig)
+	sc.pidMu.Unlock()
+}
+
+func (sc *Subcmd) terminate() error {
+	sc.pidMu.Lock()
+	if sc.current.Process == nil {
+		sc.pidMu.Unlock()
+		return ErrNotRunning
+	}
 	pid := sc.current.Process.Pid
 	sc.pidMu.Unlock()
-	syscall.Kill(-pid, syscall.SIGTERM)
-	timeout := time.After(3 * time.Second)
+	logging.Debug(sc.Log, "sending SIGINT to PID ", pid)
+	syscall.Kill(-pid, syscall.SIGINT)
+	timeout := time.After(3 * time.Second) // TODO: configurable
 	for {
 		select {
 		case <-timeout:
+			logging.Debug(sc.Log, "sending SIGKILL to PID ", pid)
 			syscall.Kill(-pid, syscall.SIGKILL)
 			return nil
 		default:
-			process, _ := os.FindProcess(pid)
+			process, _ := ps.FindProcess(pid)
 			if process == nil {
 				return nil
 			}
@@ -141,7 +171,7 @@ func (sc *Subcmd) Wait() error {
 	sc.waitMu.Lock()
 	if sc.waitCh != nil {
 		sc.waitMu.Unlock()
-		return errors.New("wait already called")
+		return ErrWaiting
 	}
 	sc.waitCh = make(chan error)
 	sc.waitMu.Unlock()
@@ -154,49 +184,63 @@ func (sc *Subcmd) setStatus(s Status) {
 		sc.statMu.Unlock()
 		return
 	}
+	logging.Debug(sc.Log, sc.status, "=>", s)
 	sc.status = s
-	sc.statMu.Unlock()
 	sc.cbMu.Lock()
 	for _, cb := range sc.callbacks {
-		cb(sc)
+		cb(sc, s)
 	}
 	sc.cbMu.Unlock()
+	sc.statMu.Unlock()
 }
 
 func (sc *Subcmd) Error() error {
+	sc.lastMu.Lock()
+	defer sc.lastMu.Unlock()
 	return sc.lastErr
 }
 
+func (sc *Subcmd) ExitStatus() int {
+	sc.lastMu.Lock()
+	defer sc.lastMu.Unlock()
+	return sc.lastStatus
+}
+
 func (sc *Subcmd) start() (err error) {
-	sc.runMu.Lock()
-	sc.setStatus(StatusStarting)
-
-	sc.pidMu.Lock()
-	sc.current = &exec.Cmd{
-		Path:        sc.Cmd.Path,
-		Args:        sc.Cmd.Args,
-		Env:         sc.Cmd.Env,
-		Dir:         sc.Cmd.Dir,
-		ExtraFiles:  sc.Cmd.ExtraFiles,
-		SysProcAttr: &syscall.SysProcAttr{Setpgid: true},
-	}
-	sc.pidMu.Unlock()
-
-	if sc.Setup != nil {
-		if err := sc.Setup(sc.current); err != nil {
-			sc.runMu.Unlock()
-			return err
-		}
-	}
-
-	err = sc.current.Start()
-	if err != nil {
-		sc.setStatus(StatusStopped)
-		sc.runMu.Unlock()
-		return err
-	}
-
+	startErr := make(chan error)
 	go func() {
+		sc.runMu.Lock()
+		sc.setStatus(StatusStarting)
+
+		sc.pidMu.Lock()
+		sc.current = &exec.Cmd{
+			Path:        sc.Cmd.Path,
+			Args:        sc.Cmd.Args,
+			Env:         sc.Cmd.Env,
+			Dir:         sc.Cmd.Dir,
+			ExtraFiles:  sc.Cmd.ExtraFiles,
+			SysProcAttr: &syscall.SysProcAttr{Setpgid: true},
+		}
+		sc.pidMu.Unlock()
+
+		if sc.Setup != nil {
+			if err := sc.Setup(sc.current); err != nil {
+				sc.runMu.Unlock()
+				startErr <- err
+				return
+			}
+		}
+
+		err = sc.current.Start()
+		if err != nil {
+			sc.setStatus(StatusStopped)
+			sc.runMu.Unlock()
+			startErr <- err
+			return
+		}
+
+		startErr <- nil
+
 		// process died too quickly?
 		if sc.current.Process == nil {
 			sc.setStatus(StatusStopped)
@@ -209,8 +253,10 @@ func (sc *Subcmd) start() (err error) {
 			sc.Started <- sc.current
 		}
 
+		sc.lastMu.Lock()
 		sc.lastErr = sc.current.Wait()
 		sc.lastStatus = exitStatus(sc.lastErr)
+		sc.lastMu.Unlock()
 		if sc.Status() != StatusStopped {
 			sc.setStatus(StatusExited)
 		}
@@ -222,26 +268,30 @@ func (sc *Subcmd) start() (err error) {
 		}
 		sc.waitMu.Unlock()
 
-		if sc.lastErr != nil && sc.lastStatus != -1 {
+		if sc.lastErr != nil && sc.lastStatus != -1 { // was not SIGKILLED
 			sc.runMu.Unlock()
 			return
 		}
 		sc.runMu.Unlock()
 
-		if sc.MaxRestarts >= 0 && sc.restarts >= sc.MaxRestarts {
+		sc.restartMu.Lock()
+		if sc.maxRestarts >= 0 && sc.restarts >= sc.maxRestarts {
+			sc.restartMu.Unlock()
 			return
 		}
+		sc.restartMu.Unlock()
 
 		if sc.Status() != StatusStopped {
 			if err := sc.start(); err != nil {
 				panic(err)
 			}
+			sc.restartMu.Lock()
 			sc.restarts++
+			sc.restartMu.Unlock()
 		}
 
 	}()
-
-	return nil
+	return <-startErr
 }
 
 func exitStatus(err error) int {
